+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this library are documented in this file.
 
 ### API Changes
 
+- Move `safe_twitter_text` to a more appropriate location in `iembot.util`.
 - Remove `purge_logs` functionality as it is better left to end users to
   implement with tooling like `tmpwatch` or log archival (#167).
 - Replace `python-twitter` dep with stdlib python code.
@@ -24,6 +25,7 @@ All notable changes to this library are documented in this file.
 - Correct Mastodon oauth authorization reset (#169).
 - Correct handling of Twitter 401s Unauthorized (#154).
 - Fix thread-safely of embedded `pymemcache` client (#148).
+- Handle Twitter HTTP 500+ errors more gracefully (#171).
 - Implement ATmosphere/bluesky media upload again (#144).
 - Implement webhooks route registration (#165).
 - Improve mastodon coverage and error handling (#161).